Move MainTimer component to components/Elements dir
